+++ packages/gov-ui-kit/src/core/components/dropdown/index.ts
@@ -1,10 +1,10 @@
 import { DropdownContainer } from './dropdownContainer';
 import { DropdownItem } from './dropdownItem';
 
-export type { IDropdownContainerProps } from './dropdownContainer';
-export type { IDropdownItemProps } from './dropdownItem';
-
 export const Dropdown = {
     Container: DropdownContainer,
     Item: DropdownItem,
 };
+
+export * from './dropdownContainer';
+export * from './dropdownItem';